+++ RELEASE_NOTES.md
@@ -35,6 +35,7 @@ A new header is inserted each time a *tag* is created.
   light's intensity in candela.
 - Fixed an issue where some `ShadowOptions` were not being respected when passed to
   `LightManager::Builder`.
+- Added a Depth of Field post-processing effect
 
 ## v1.5.2